+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ History
 Next Release
 ------------
 * Feat: Add support for grouping elements (#72)
+* Feat: Locking workspace in ``with`` block through ``lock()`` method (#62)
 
 0.4.0 (2021-02-05)
 ------------------